requirements.txt
django==5.1.4
djangorestframework==3.15.2
gunicorn==23.0.0
whitenoise==6.8.2
reportlab==4.2.5
